backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Darshit Kothari <kotharidarsh104@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.11.3"
fastapi = "^0.54.1"
fastapi-mail = "^1.2.4"
h11 = "^0.14.0"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.4"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^3.1.2"
PyJWT = "^2.6.0"
PyYaml = "6.0"
alembic = "^1.4.2"
sqlalchemy = "^1.4.46"
pytest = "^5.4.1"
pytest-mock = "^3.6.1"
aiounittest = "^1.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
starlette-exporter = "^0.4.1"
starlette-context = "^0.2.3"
aiofiles = "^0.5.0"
cryptography = "3.3.2"
aiohttp = "^3.7.4"
pika = "^1.2.0"
APScheduler = "^3.7.0"
pytz = "^2021.1"
httplib2 = "^0.20.2"
celery = {version = "^5.1.2", extras = ["redis"]}
redis = "^4.4.2"
urllib3 = "^1.26.14"
websockets = "10.4"
mysql-connector-python = "^8.0.32"
pymysql = "^1.0.2"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
watchdog = "^2.1.3"
argh = "^0.26.2"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
